outputs.postgresql: remove dbConnectedHook for clearing caches
Remove the code that cleared caches on total DB connection loss. The intent was to be able to reset known state in the event of things like a database failover. But this is such an edge case, it's not worth the maintenance of the code.
If we ever want it back, it's here, in the git history.